refactor: rename Tokenizer fields, move types to more logical places, and stop keeping track of the current line in the Tokenizer

+++ src/Diagnostics.zig
@@ -6,7 +6,7 @@ const Diagnostics = @This();
 
 const std = @import("std");
 
-line_number: u64 = 1,
+line_number: usize = 1,
 column: usize = 0,
 snippet: []const u8 = "",
 message: []const u8 = "",
